Admin Section
Added an admin section that can only be accessed by users with
"user_type" of 99 or more. If an admin logs in, he/she will
automatically be redirected to this page.

The page also checks your user type on load dynamically. If you do not
belong on the page then it spits you back to index.jsp (keeps out the
"technical savvy" people who like to poke around).

The admin section makes use of jQuery.accordion() to display all of the
active suggestions. From there, the admin has the ability to either
accept or reject the suggestion.

On accept:
Post moved to active_posts, archived_posts
On reject:
Post moved to archived_posts_rejected